WIP: Fix Ogmo Entity/Decal offset positions
- Change the game into a testing ground (convert to test before merging)
- Reverse Ogmo level layer order so the draw order is consistent with how things are rendered in Ogmo Editor
- Add Geometry component to each level layer Entity so its size is properly defined and it can provide a relative origin offset for its children
- Properly convert Entity and Decal positions such that custom origin in Ogmo is supported, and positions line up correctly with tile layers

TODO: rotation is always assumed to be about the center of an entity, but the origin for an entity in Ogmo Editor defaults to top-left. We should properly handle the rotation origin so that rotated entities will have the same apparent location in Hex Engine as in Ogmo Editor (right now that's only true if the entity origin is changed in Ogmo Editor to not be the top-left). However, this may require changes to Geometry, because rotation origin has up to this point always been assumed to be the same as position origin.